XWIKI-17524: Replace Jobs with threads to handle mentions asynchronously
- various fixes and defensive code added
- improved internal APIs

+++ xwiki-platform-core/xwiki-platform-mentions/xwiki-platform-mentions-default/src/main/java/org/xwiki/mentions/internal/jmx/JMXMentionsMBean.java
@@ -31,7 +31,7 @@ public interface JMXMentionsMBean
      *
      * @return the mentions analysis tasks queue size.
      */
-    int getQueueSize();
+    long getQueueSize();
 
     /**
      * Clear the mentions analysis tasks queue.
